provide basic bindings for macOS frameworks
This change adds a work-in-progress support foreign function calls, including
a working callback example, though the latter would be tricky to implement as
part of the new cabi package.

Most of the functionality is now focused on providing bindings for macOS APIs.

The dyld package is internal, but it’ll exposed as a standalone Go module some
time in the future.

The main limitation of the current implementation is that there seems to be no
way to detect that Go is being compiled in external linker mode, and in that
mode cgo_import_dynamic directives are no-op.

This means that the user must explicitly link against the required libraries
when using the external linker. That said, macOS is a first-class target in
Go and internal linker works just fine, so there should be no reason to use
external linker.